Shared the Hub XML-RPC API setup between install and migrate
Migrate also needs to setup the Hub XML-RPC API container. Reshuffling
the code a bit to share what install is doing.
