+++ mgradm/shared/utils/types.go
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: 2024 SUSE LLC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package utils
+
+import "github.com/uyuni-project/uyuni-tools/shared/types"
+
+// HubXmlrpcFlags contains settings for Hub XMLRPC container.
+type HubXmlrpcFlags struct {
+	Replicas int
+	Image    types.ImageFlags `mapstructure:",squash"`
+}
